Create index.html
This code will create a navigation bar at the right hand of the page with the following links:

    Home
    Products
    Services
    About
    Contact

The navigation bar will be purple in color with lemon green text. When a user hovers over a link, the text will change to white. The rest of the page will be white with a lemon green header.