Create LIVE.html
This code will create a live chat that is displayed in a fixed position at the bottom right of the screen. Users can join the chat by typing a message into the input field and clicking the "Send" button. Messages from other users will be displayed in a list below the input field. Users can also react to messages by clicking the reaction icons.