Create Terms-And-Privacy.html
This code will add a new tab to the navigation bar called "Terms and Privacy". When a user clicks on this tab, they will be shown a page with the terms and conditions of using the web application. The terms and conditions can be formatted however you like, but the code above provides a basic example.