Create change-color-theme.html
HTML code to change the webapp's color theme                                             This code uses a CSS variable called :root to define the default colors for the webapp. The :root selector is used to target the highest-level element in the DOM, which in this case is the html element. The :root selector can be used to define the default colors for the entire webapp.