Create search-bar,-menu-tab,-and-SMS-tab.html
This code will add a search bar, a menu tab, and an SMS tab to the web application. The search bar will allow users to search for other users, pages, groups, and products. The menu tab will contain links to the different pages of the web application. The SMS tab will allow users to send and receive SMS messages.