Create MTN-and-Orange-Momo-Payment-Services.html
This code will add two payment buttons to the webapp, one for MTN Mobile Money and one for Orange Mobile Money. When a user clicks on a payment button, they will be taken to a secure payment page where they can enter their phone number and the amount they want to pay. Once the payment is made, the user will be returned to the webapp and their purchase will be completed.

To use this code, you will need to create a payment account with MTN Mobile Money or Orange Mobile Money. You will also need to get the API credentials for your payment account. Once you have these credentials, you can replace the placeholder values in the code with your own values.