Create contact-travel-agency.html
This code will create a contact form that users can use to contact a travel agency. The form will collect the user's name, email address, phone number, travel agency, and message. The form will then send the information to the travel agency's website.

The code also includes the MTN Mobile Money and Orange Mobile Money payment code. These codes allow users to pay for their purchases with MTN Mobile Money or Orange Mobile Money.

To use this code, you will need to modify the action attribute of the form to point to the website of the travel agency that you want to contact. You will also need to modify the name and placeholder attributes of the form fields to match the names and placeholders that you want to use.