Create dashboard1.html
simple HTML, CSS, and JavaScript code that creates a basic e-commerce social media webapp. The code defines the following elements:

    A navigation bar with links to different pages on the website.
    A header with the title of the website.
    A post with a title, body, and comments.
    A section with posts from other users.
    A section with uploads from other users.

The code also uses the JavaScript library Firebase to get the username of the current user from the database. This username is then used to set the value of the .post-comment-author element in the DOM.