Create firebase-signup.html
This code will first initialize Firebase and then add an event listener to the signup button. When the button is clicked, the email and password from the form will be retrieved and used to create a new user with Firebase. If the user is created successfully, the user will be redirected to the login page.