Create firebase_-login_rdr_dashboard2.html
This code will first initialize Firebase and then add an event listener to the login button. When the button is clicked, the username and password from the login form will be retrieved. These values will then be used to check if they match the signup details stored in the Firebase database. If they do match, the user will be redirected to the dashboard.html page. If they do not match, an error message will be displayed.