Create style.css
This CSS code will make the tabs squircle-shaped, stack them on top of each other, and center them vertically. It will also change the color of the tabs when they are hovered over.

To use this CSS code, you would need to add it to the <head> section of your HTML document. For example:
HTML

<head>
  <title>My Website</title>
  <link rel="stylesheet" href="style.css">
</head>



The style.css file would then contain the CSS code that you provided.